# proxy/src/docs-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Sed vel mollis lectus, quis rutrum augue. Etiam mattis volutpat hendrerit. Suspendisse nulla ex, tincidunt ac metus sed, tempor ultrices dolor. Integer eget egestas diam, blandit finibus nisl. Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Suspendisse potenti. Maecenas sit amet turpis iaculis, fringilla quam vitae, mattis erat. Nam ut ullamcorper ligula. Nulla auctor semper dignissim. Pellentesque nec eros hendrerit dolor scelerisque porttitor non sed sapien. Nullam aliquam bibendum quam vel vulputate. Nunc non mi risus.",
    "version": "1.0.0",
    "title": "API Documentation",
    "contact": { "email": "ikballukmanulhakim5@gmail.com" }
  },
  "servers": [
    {
      "url": "http://localhost:5000/",
      "description": "Local Server"
    }
  ],
...
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login",
        "description": "Login untuk mengakses API",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "username": {
                    "type": "string",
                    "example": "user1"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "required": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "example": {
                  "data": {
                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImVtYWlsMkBtYWlsLmNvbSIsInVzZXJuYW1lIjoidXNlcjIiLCJ1c2VySWQiOiJlYzE0YjExZS05MTU1LTQ0NDEtYjlhNy0zNjFmYTRiZTJmNWQiLCJpYXQiOjE2Mzg1MTk4NTYsImV4cCI6MTYzODU0ODY1Nn0.F1HiKhvweR8QK2npzbqi-KipV_Rx_ibcnhnTfYmGTsw"
                  }
                }
              }
            }
          },
          "400": {
            "description": "error authorized failed",
            "content": {
              "application/json": {
                "example": {
                  "message": "password required"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
      }
    }
  }
}